Make date range accept start and stop times
Sometimes you want to have more granular control over the allowed
window when a code is active.

For instance, you want to hand out a single use code, but it can only be
used on a certain day two weeks in the future during a given set of
hours.

Or, you have a guest coming that you want to give them an access code,
but you know that they won't be there before say, 13:00 on a given day
and they'll be leaving by 15:00 on a day in the following week.

In any of these cases you don't want the code to be active all day on
either of the days that they are arriving or leaving, but you also can't
just use a date range and limit the hours using the more granular by day
options.

Fixes: #144
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>