Feat: Advanced options toggles (#461)
* Feat: Add new schema for advanced features

Add the schema and migration for being able to disable the advanced
features of date range and day of week. As the features have
historically been enabled by default, this migration will set the
default values to True for existing installations.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>

* Feat: Teach config flow about advanced options

Teaches the config flow about the advanced options for date range and
day of week.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>

* Feat: Teach how to skip Advanced options

Teach Keymaster how to skip configuration of the entities for the
Advanced Date Range and Advanced Day of Week options.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>

* Feat: Teach dashboard generator advanced toggles

Teach the dashboard generator about the advanced toggles for date range
and day of week. When a customer configures keymaster with these options
the related entities will be appropriately added to generated dashboard
output.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>

---------

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>